Implemented the remaining IP Tunneling protocols - IP6over4, IP4over6, IP6over6; renumbered L3 protocol protobuf "field numbers". ComboProtocol now implements PseudoIp cksum as protoB's cksum as a convenience for the IP tunneling protocols. Also found and fixed an issue with IPv6's pseudoIp cksum which was causing the TCP/UDP cksums to be incorrect

+++ common/ip6.cpp
@@ -819,13 +819,13 @@ quint32 Ip6Protocol::protocolFrameCksum(int streamIndex,
                 .toByteArray();
         Q_ASSERT(addr.size() == 16);
         for (int i = 0; i < addr.size(); i+=2)
-            sum += (addr.at(i) << 8) + addr.at(i+1);
+            sum += (quint8(addr.at(i)) << 8) + quint8(addr.at(i+1));
 
         addr = fieldData(ip6_dstAddress, FieldFrameValue, streamIndex)
                 .toByteArray();
         Q_ASSERT(addr.size() == 16);
         for (int i = 0; i < addr.size(); i+=2)
-            sum += (addr.at(i) << 8) + addr.at(i+1);
+            sum += (quint8(addr.at(i)) << 8) + quint8(addr.at(i+1));
 
         sum += fieldData(ip6_payloadLength, FieldValue, streamIndex)
                 .toUInt() & 0xFFFF;
